Rename player to playerSlice
