+++ components/player/PlayerController.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled'
 import { RootState } from 'features'
-import { playerSelectors, PlayerTime } from 'features/player'
+import { playerSelectors, PlayerTime } from 'features/playerSlice'
 import React from 'react'
 import { useSelector } from 'react-redux'
