+++ features/index.ts
@@ -2,8 +2,8 @@ import { combineReducers } from '@reduxjs/toolkit'
 import { configureStore } from '@reduxjs/toolkit'
 import { combineEpics, createEpicMiddleware } from 'redux-observable'
 
-import { playerReducer } from './player'
 import { playerFetchEpic, playerRewindEpic, playerToggleEpic } from './playerEpic'
+import { playerReducer } from './playerSlice'
 
 const rootReducer = combineReducers({
   player: playerReducer,
